Finished Introduction Section
Spent some time relearning material and even more time learning more material. Learned trying to brute force my sections & divs was not working, so I switched gears and committed to learning flexboxes. The result is much cleaner and easy to use and understand. As of now, the introduction section is presentable; now to continue work on the next sections.