Uploading Portraits & Minor Visual Changes
Quickly updated the About Me and Contact Me sections to have similar style to the completed Introduction section in terms of font sizes and colors. Also went ahead and uploaded some self portraits and pictures for the About Me and Contact Me sections to be completed at a later time.