Updated Contact Me Section, Added Project Screenshots, & Organized Images
In this update, I styled my three existing social media divs in the Contact Me section to be more appealing to users. I also added screenshots and a video to almost all of the listed projects in the My Projects section. Did a little bit of organizing to more easily find some images in the Images folder. Almost finished and presentable; will need to come back and polish code at a later time.