Added Strava link and Text in About Me Section
Went ahead and added my Strava in the Contact Me section of the site, and I made the titles of the socials in the divs links to their corresponding websites (clicking the general div doesn't do anything; may have to switch to buttons at a later time). Also finally added some text the About Me section of the site. Finally, I was able to center the screenshots and videos of the listed projects in the My Projects section. Next step is code clean up.